added basic admin features, sanitized all inputs, various other improvements

+++ routes/index.js
@@ -2,6 +2,7 @@ var express = require('express');
 var passport = require('passport');
 var Account = require('../models/account');
 var	mongoClient = require('mongodb').MongoClient;
+var sanitize = require('mongo-sanitize');
 var router = express.Router();
 
 var siteData = require('../site-data').siteData;
@@ -40,64 +41,69 @@ router.get('/region/:regionid', function( req, res) {
 
 // Retrieve dialog
 router.get('/dialogs/:filename', function (req, res) {
-	res.render(req.originalUrl.substr(1), {
-		regionIcons: siteData.dialog.regionIcons,
-		regionMarkers: siteData.dialog.regionMarkerShapes
-		});
+	if (req.isAuthenticated()) {
+		res.render(req.originalUrl.substr(1), {
+			regionIcons: siteData.dialog.regionIcons,
+			regionMarkers: siteData.dialog.regionMarkerShapes
+			});
+	}
 });
 
 // Add a new sound marker
 router.post('/submit', function(req, res) {
-	var doc = {
-		'lat': req.body.lat,
-		'lng': req.body.lng,
-		'region': req.body.region,
-		'date': req.body.date,
-		'type': req.body.type,
-		'media': req.files[0].filename,
-		'creator': req.user.username,
-		'tags': []
-	};
-	markerCollection.update(
-		{ regionName: req.body.region},
-		{
-			$push: { markers: doc }
-		},
-		{ upsert: true }
-	);
-	console.log('Added new marker');
-	res.end('SUCCESS');
+	if (req.isAuthenticated()) {
+		var doc = {
+			'lat': req.body.lat,
+			'lng': req.body.lng,
+			'region': req.body.region,
+			'date': req.body.date,
+			'type': req.body.type,
+			'media': req.files[0].filename,
+			'creator': req.user.username,
+			'tags': []
+		};
+		markerCollection.update(
+			{ regionName: sanitize(req.body.region)},
+			{
+				$push: { markers: doc }
+			},
+			{ upsert: true }
+		);
+		console.log('Added new marker');
+		res.end('SUCCESS');
+	}
 });
 
 router.post('/user', function(req, res) {
-	if (req.user) {
-		accounts.update({ _id: req.user._id}, { $set: {image: req.files[0].filename}});
+	if (req.isAuthenticated()) {
+		accounts.update({ _id: sanitize(req.user._id)}, { $set: {image: sanitize(req.files[0].filename)}});
 		res.end('SUCCESS');
 	};
 });
 
 router.post('/update_tags', function (req, res) {
-	var tags = JSON.parse(req.body.tags);
-	console.log(req.body.tags);
-	markerCollection.update(
-		{ 'regionName': req.body.region,
-			'markers.media': req.body.media },
-		{
-			$set: { 'markers.$.tags': tags}
-		}
-	);
-	res.end('SUCCESS');
-	console.log('updated tags on '+req.body.media);
+	if (req.isAuthenticated()) {
+		var tags = JSON.parse(req.body.tags);
+		markerCollection.update(
+			{ 'regionName': sanitize(req.body.region),
+				'markers.media': sanitize(req.body.media) },
+			{
+				$set: { 'markers.$.tags': tags}
+			}
+		);
+		res.end('SUCCESS');
+		console.log('updated tags on '+req.body.media);
+	}
 });
 
 router.post('/delete_marker', function (req, res) {
-	if (req.user) {
+	if (req.isAuthenticated()) {
 		markerCollection.update(
-			{ 'regionName': req.body.region,
-				'markers.media': req.body.media,
-			 	'markers.creator': req.user.username },
+			{ 'regionName': sanitize(req.body.region),
+				'markers.media': sanitize(req.body.media),
+			 	'markers.creator': sanitize(req.user.username) },
 			{
-				$pull: { 'markers': { media: req.body.media}}
+				$pull: { 'markers': { media: sanitize(req.body.media)}}
 			}
 		);
 		res.end('SUCCESS');
@@ -109,32 +115,35 @@ router.post('/delete_marker', function (req, res) {
 });
 
 router.post('/update_marker_order', function(req, res) {
-	console.log(JSON.parse(req.body.markers));
-	console.log(markerCollection.update(
-		{ '_id': req.body.regionId },
-		{
-			$set: {'markers' : JSON.parse(req.body.markers)}
-		}
-	));
-	res.end('SUCCESS');
+	if (req.isAuthenticated()) {
+		console.log(markerCollection.update(
+			{ '_id': sanitize(req.body.regionId) },
+			{
+				$set: {'markers' : JSON.parse(sanitize(req.body.markers))}
+			}
+		));
+		res.end('SUCCESS');
+	}
 });
 
 // Add a new sound region
 router.post('/submit_region', function(req, res) {
-	var region = {
-		'regionName': req.body.regionName,
-		'lat': req.body.lat,
-		'lng': req.body.lng,
-		'color': req.body.color,
-		'icon': req.body.icon,
-		'shape': req.body.shape,
-		'markers': [],
-		'geofence': req.body.geofence,
-		'type': req.body.type
-	};
-	markerCollection.insert(region);
-	res.end('SUCCESS');
-	console.log('Added new region');
+	if (req.isAuthenticated()) {
+		var region = {
+			'regionName': sanitize(req.body.regionName),
+			'lat': sanitize(req.body.lat),
+			'lng': sanitize(req.body.lng),
+			'color': sanitize(req.body.color),
+			'icon': sanitize(req.body.icon),
+			'shape': sanitize(req.body.shape),
+			'markers': [],
+			'geofence': sanitize(req.body.geofence),
+			'type': sanitize(req.body.type)
+		};
+		markerCollection.insert(region);
+		res.end('SUCCESS');
+		console.log('Added new region');
+	}
 });
 
 // retrieve markers
@@ -142,26 +151,32 @@ router.get('/get_markers', function(req, res) {
 	markerCollection.find().toArray( function(err, items) {
 		res.send(JSON.stringify(items));
 	});
-
 	console.log('Sending markers');
 });
 
 // Wipe database (Temporary)
 router.post('/self_destruct', function(req, res) {
-	console.log('Self destructing');
-	markerCollection.drop();
-	accounts.drop();
+	if (req.isAuthenticated() && req.user.lvl == 'admin') {
+		console.log('Self destructing');
+		markerCollection.drop();
+		//accounts.drop();
+	}
 });
 
+router.get('/admin', function (req, res) {
+	if (req.isAuthenticated() && req.user.lvl == 'admin') {
+		res.render('admin.pug', { user: req.user});
+	}
+})
 
 router.get('/register', function(req, res) {
 	res.render('register.pug', { } );
 });
 
-router.get('/username_available', function(req, res) {
-	accounts.findOne({ 'username': req.body.username}).then((e) => {
+router.post('/username_available', function(req, res) {
+	accounts.findOne( { 'username': sanitize(req.body.query) }).then((e) => {
 		if (e != null) {
-			res.status(400).end({ error: 'username taken'});
+			res.status(401).end('ERROR');
 		} else {
 			res.status(200).end('SUCCESS');
 		}
@@ -169,10 +184,15 @@ router.get('/username_available', function(req, res) {
 });
 
 router.post('/register', function(req, res) {
+	var userLvl = 'user';
+	if (req.body.username == 'admin') {
+		userLvl = 'admin';
+	}
 	Account.register(new Account({
-		email: req.body.email,
-		username: req.body.username,
-		name: req.body.name,
+		email: sanitize(req.body.email),
+		username: sanitize(req.body.username),
+		name: sanitize(req.body.name),
+		lvl: userLvl
 	}), req.body.password, function(err, account) {
 		if (err) {
 			console.log(err);
@@ -189,12 +209,13 @@ router.get('/login', function(req, res) {
 	res.render('login.pug', {user: req.user });
 });
 
-router.post('/login', passport.authenticate('local'), function(req, res) {
-	res.redirect('/');
-});
+router.post('/login', passport.authenticate('local', {
+	successRedirect: '/',
+	failureRedirect: '/login'
+}));
 
 router.get('/user', function(req, res) {
-	if (req.user) {
+	if (req.isAuthenticated()) {
 		res.render('user.pug', {user: req.user });
 	} else {
 		res.redirect('/login');
@@ -202,7 +223,7 @@ router.get('/user', function(req, res) {
 });
 
 router.get('/get_user_markers', function(req, res) {
-	if (req.user) {
+	if (req.isAuthenticated()) {
 		markerCollection.aggregate([
 				{
 					$project: {
@@ -210,7 +231,7 @@ router.get('/get_user_markers', function(req, res) {
 							$filter: {
 								input: "$markers",
 								as: "marker",
-								cond: { $eq: ["$$marker.creator", req.user.username]}
+								cond: { $eq: ["$$marker.creator", sanitize(req.user.username)]}
 							}
 						}
 					}
